requirements.txt
Flask
openai
requests
flask-sqlalchemy
unbabel-comet